+++ routes/web.php
@@ -2,11 +2,12 @@
 
 use Illuminate\Support\Facades\Route;
 use Rahmanramsi\LivewirePageGroup\Facades\LivewirePageGroup;
+use Rahmanramsi\LivewirePageGroup\PageGroup;
 
 Route::name('livewirePageGroup.')
     ->group(function () {
         foreach (LivewirePageGroup::getPageGroups() as $pageGroup) {
-            /** @var \Rahmanramsi\LivewirePageGroup\PageGroup $pageGroup */
+            /** @var PageGroup $pageGroup */
             $pageGroupId = $pageGroup->getId();
 
             Route::domain($pageGroup->getDomain())