+++ src/PageGroupServiceProvider.php
@@ -9,7 +9,7 @@ abstract class PageGroupServiceProvider extends ServiceProvider
 {
     abstract public function pageGroup(PageGroup $pageGroup): PageGroup;
 
-    public function register()
+    public function register(): void
     {
         $this->app->resolving('livewire-page-group', function () {
             LivewirePageGroup::registerPageGroup(