+++ tests/Fixtures/views/layouts/app.blade.php
@@ -1,0 +1,10 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <title>{{ $livewire->getTitle() }}</title>
+</head>
+<body>
+    <header>Livewire Page Group test layout</header>
+    {{ $slot }}
+</body>
+</html>